TASK: Detect and render required emptyHiddenFields
The required empty hidden fields for checkboxes and multiselects are detected via xPath. The emptyHiddenFieldnames method
of the Form helper returns an array of the required fieldnames which is then rendered in the form prototype.

In addition the xPath for the trustedProperties was adjusted only match form objects and respect the fieldnamePrefix.