TASK: Add separation of Field and FieldComponent
The Field is now a container with the sole purpose of determining the name, submitted or bound values and validation results. Thus it only supports
`name` and `property` as arguments (form is taken from context) It provides the elements below with the `field` context to read name, value and validationResults from.

The FieldComponent is a base prototype for the actual inputs it supports `id`, `class`, `required`, `attributes`. The purpose of the field is to render the actual html-control
it uses existing field definition in the context but will create a local one if `name` or `property` are found so The elements can also be used directly without Field Container.