FEATURE: Add generic `data`-binding instead of single `name` + `object` binding
The form attributes `name` and `object` are replaced by a single `data` attribute to allow the
mapping of multipleObjects and also simple values.

Note: As result all field `property` and `name` attributes have to include the object name now.

The example shows a form that maps two objects `customer` and `shippingAddress` that
are submitted to the targetAction:
```
<Neos.Fusion:Form
   data.customer={customer}
   data.shippingAddress={shippingAddress}
>
             <Neos.Fusion.Form:Textfield property="customer.firstName" />
             <Neos.Fusion.Form:Textfield property="customer.lastName" />

             <Neos.Fusion.Form:Textfield property="shippingAddress.street" />
             <Neos.Fusion.Form:Textfield property="shippingAddress.city" />

	     <Neos.Fusion.Form:Submit value="submit />
</Neos.Fusion:Form>
```