BUGFIX: Handle selecting multiple entities properly
The trusted properties have to be aware of every checkbox and select option for multivalue selects or checkboxes
to correctly calculate the number of maximal allowed items for property mapping.

This fix removes the calling of array_unique for the fieldnames that are used to generate the trusted properties.
For multi selects the name is added for every option so the maximal possible number of items can be mapped.