!!! TASK: Definition of form `namespace`
The key `namespacePrefix` from the documentation did not match the `fieldNamePrefix` that was read nor `fieldnamePrefix` as was defined.

How to upgrade:  If you perviously used a `fieldNamePrefix` for your FusionForm you have to change this key to `namespace` now.
This is quite unlikely as this never worked as documented and was seldom used anyways.