BUGFIX: Render bound data when submitted values are null
When validation errors were present the field value only tries to access the submitted values.
This caused confusion when bound data had a value for the path but the submitted values did not.

The change adds a fallback from submitted values to bound data in case the submitted data is null.

Resolves: #48