BUGFIX: Allow back navigation on confirmation step.
A Multistep process is considered to be finished once all steps are committed. This causes problems on the confirmation step as
this adds no data and thus is always considered valid once submitted. Because of that the process was finished even after the user
navigated back on the confirmation step.

This change alters the multiStepProcess to only declare itself finished if no target step is set.