BUGFIX: Make previous data available to the fusion contexts before subprocesses are evaluated.
Currently the `data` is only pushed to the context before the subprocess is rendered. This change ensures
that the data ins also in the context before the subprocesses of the multiStepForm are evaluated. That way
schemas may use the data for instance to configure fields as required if other fields are set.

```
schema {
    conditionalField = ${Form.Schema.string()}
    conditionalField.@process.makeRequired = ${value.isRequired()}
    conditionalField.@process.makeRequired.@if.hasOtherValue = ${data.otherValue}
}
```

This even allows to make whole steps conditional

```
process = Neos.Fusion.Form:Runtime.MultiStepProcess {
    steps {
        address {
            ...
        }
        visa {
            @if.fromForeignGalaxy = ${data.galaxy != 'milkyway'}
            ...
        }
    }
}
```